switch from attributeNames to attributeIds
In order to implement PATCH requests and Patch entityUpdates,
we want to switch to attributeIds in the TypeModels, as well as in the
JSON bodies for server requests and responses.

This commit migrates to attributeIds for all places except the
OfflineStorage and TypeRefs. Later, we also want to migrate to
attributeIds in the OfflineStorage, enabling to run JSON MERGE Patch,
and by that remove all needs for backward migrations on server
responses and all offline migrations.

Co-authored-by: Kinan <104761667+kibibytium@users.noreply.github.com>