Fix delete action for conversation to only delete mails in trash/spam
When permanently deleting a conversation that has some of its mails in
folders that aren't in trash/spam, only mails that are in the current
folder which must be trash/spam are deleted.
This is done by passing the folder id to the DeleteMailService along
with all the conversation mailIds.

Close #8658

Co-authored-by: bir <bir@tutao.de>